+++ platforms/software/dependency-management/src/main/java/org/gradle/api/internal/attributes/DefaultAttributesSchema.java
@@ -16,34 +16,26 @@
 
 package org.gradle.api.internal.attributes;
 
-import com.google.common.base.Objects;
 import org.gradle.api.Action;
 import org.gradle.api.attributes.Attribute;
 import org.gradle.api.attributes.AttributeMatchingStrategy;
+import org.gradle.api.internal.attributes.matching.AttributeMatcher;
+import org.gradle.api.internal.attributes.matching.CachingAttributeSelectionSchema;
+import org.gradle.api.internal.attributes.matching.DefaultAttributeMatcher;
+import org.gradle.api.internal.attributes.matching.DefaultAttributeSelectionSchema;
 import org.gradle.internal.Cast;
-import org.gradle.internal.component.model.AttributeMatcher;
-import org.gradle.internal.component.model.AttributeSelectionSchema;
-import org.gradle.internal.component.model.AttributeSelectionUtils;
-import org.gradle.internal.component.model.DefaultAttributeMatcher;
-import org.gradle.internal.component.model.DefaultCompatibilityCheckResult;
-import org.gradle.internal.component.model.DefaultMultipleCandidateResult;
 import org.gradle.internal.instantiation.InstantiatorFactory;
 import org.gradle.internal.isolation.IsolatableFactory;
 
 import javax.annotation.Nullable;
 import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collection;
 import java.util.Collections;
 import java.util.HashMap;
-import java.util.LinkedHashMap;
 import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
-import java.util.stream.Collectors;
-import java.util.stream.IntStream;
 
 public class DefaultAttributesSchema implements AttributesSchemaInternal {
     private final InstantiatorFactory instantiatorFactory;
@@ -101,7 +93,12 @@ public class DefaultAttributesSchema implements AttributesSchemaInternal {
     @Override
     public AttributeMatcher withProducer(AttributesSchemaInternal producerSchema) {
         return matcherCache.computeIfAbsent(producerSchema, key ->
-            new DefaultAttributeMatcher(new DefaultAttributeSelectionSchema(this, producerSchema)));
+            new DefaultAttributeMatcher(
+                new CachingAttributeSelectionSchema(
+                    new DefaultAttributeSelectionSchema(this, producerSchema)
+                )
+            )
+        );
     }
 
     @Override
@@ -151,207 +148,6 @@ public class DefaultAttributesSchema implements AttributesSchemaInternal {
     @Override
     public Attribute<?> getAttributeByName(String name) {
         return attributesByName.get(name);
-    }
-
-    // TODO: Move this out into its own class so it can be unit tested directly.
-    private static class DefaultAttributeSelectionSchema implements AttributeSelectionSchema {
-        private final AttributesSchemaInternal consumerSchema;
-        private final AttributesSchemaInternal producerSchema;
-
-        private final Map<ExtraAttributesEntry, Attribute<?>[]> extraAttributesCache;
-
-        public DefaultAttributeSelectionSchema(AttributesSchemaInternal consumerSchema, AttributesSchemaInternal producerSchema) {
-            this.consumerSchema = consumerSchema;
-            this.producerSchema = producerSchema;
-            this.extraAttributesCache = new ConcurrentHashMap<>();
-        }
-
-        @Override
-        public boolean hasAttribute(Attribute<?> attribute) {
-            return consumerSchema.getAttributes().contains(attribute) || producerSchema.getAttributes().contains(attribute);
-        }
-
-        @Override
-        public Set<Object> disambiguate(Attribute<?> attribute, @Nullable Object requested, Set<Object> candidates) {
-            DefaultMultipleCandidateResult<Object> result = null;
-
-            DisambiguationRule<Object> rules = consumerSchema.disambiguationRules(attribute);
-            if (rules.doesSomething()) {
-                result = new DefaultMultipleCandidateResult<>(requested, candidates);
-                rules.execute(result);
-                if (result.hasResult()) {
-                    return result.getMatches();
-                }
-            }
-
-            rules = producerSchema.disambiguationRules(attribute);
-            if (rules.doesSomething()) {
-                if (result == null) {
-                    result = new DefaultMultipleCandidateResult<>(requested, candidates);
-                }
-                rules.execute(result);
-                if (result.hasResult()) {
-                    return result.getMatches();
-                }
-            }
-
-            if (requested != null && candidates.contains(requested)) {
-                return Collections.singleton(requested);
-            }
-
-            return null;
-        }
-
-        @Override
-        public boolean matchValue(Attribute<?> attribute, Object requested, Object candidate) {
-            if (requested.equals(candidate)) {
-                return true;
-            }
-
-            CompatibilityCheckResult<Object> result = null;
-
-            CompatibilityRule<Object> rules = consumerSchema.compatibilityRules(attribute);
-            if (rules.doesSomething()) {
-                result = new DefaultCompatibilityCheckResult<>(requested, candidate);
-                rules.execute(result);
-                if (result.hasResult()) {
-                    return result.isCompatible();
-                }
-            }
-
-            rules = producerSchema.compatibilityRules(attribute);
-            if (rules.doesSomething()) {
-                if (result == null) {
-                    result = new DefaultCompatibilityCheckResult<>(requested, candidate);
-                }
-                rules.execute(result);
-                if (result.hasResult()) {
-                    return result.isCompatible();
-                }
-            }
-
-            return false;
-        }
-
-        @Override
-        public Attribute<?> getAttribute(String name) {
-            Attribute<?> attribute = consumerSchema.getAttributeByName(name);
-            if (attribute != null) {
-                return attribute;
-            }
-            return producerSchema.getAttributeByName(name);
-        }
-
-        @Override
-        public Attribute<?>[] collectExtraAttributes(ImmutableAttributes[] candidateAttributeSets, ImmutableAttributes requested) {
-            // It's almost always the same attribute sets which are compared, so in order to avoid a lot of memory allocation
-            // during computation of the intersection, we cache the result here.
-            ExtraAttributesEntry entry = new ExtraAttributesEntry(candidateAttributeSets, requested);
-            return extraAttributesCache.computeIfAbsent(entry, key ->
-                AttributeSelectionUtils.collectExtraAttributes(this, candidateAttributeSets, requested));
-        }
-
-        @Override
-        public PrecedenceResult orderByPrecedence(Collection<Attribute<?>> requested) {
-            if (consumerSchema.getAttributeDisambiguationPrecedence().isEmpty() && producerSchema.getAttributeDisambiguationPrecedence().isEmpty()) {
-                // If no attribute precedence has been set anywhere, we can just iterate in order
-                return new PrecedenceResult(IntStream.range(0, requested.size()).boxed().collect(Collectors.toList()));
-            } else {
-                // Populate requested attribute -> position in requested attribute list
-                final Map<String, Integer> remaining = new LinkedHashMap<>();
-                int position = 0;
-                for (Attribute<?> requestedAttribute : requested) {
-                    remaining.put(requestedAttribute.getName(), position++);
-                }
-                List<Integer> sorted = new ArrayList<>(remaining.size());
-
-                // Add attribute index to sorted in the order of precedence by the consumer
-                for (Attribute<?> preferredAttribute : consumerSchema.getAttributeDisambiguationPrecedence()) {
-                    if (requested.contains(preferredAttribute)) {
-                        sorted.add(remaining.remove(preferredAttribute.getName()));
-                    }
-                }
-                // Add attribute index to sorted in the order of precedence by the producer
-                for (Attribute<?> preferredAttribute : producerSchema.getAttributeDisambiguationPrecedence()) {
-                    if (remaining.containsKey(preferredAttribute.getName()) && requested.contains(preferredAttribute)) {
-                        sorted.add(remaining.remove(preferredAttribute.getName()));
-                    }
-                }
-                // If nothing was sorted, there were no attributes in the request that matched any attribute precedences
-                if (sorted.isEmpty()) {
-                    // Iterate in order
-                    return new PrecedenceResult(remaining.values());
-                } else {
-                    // sorted now contains any requested attribute indices in the order they appear in
-                    // the consumer and producer's attribute precedences
-                    return new PrecedenceResult(sorted, remaining.values());
-                }
-            }
-        }
-
-        @Override
-        public boolean equals(Object o) {
-            if (this == o) {
-                return true;
-            }
-            if (o == null || getClass() != o.getClass()) {
-                return false;
-            }
-            DefaultAttributeSelectionSchema that = (DefaultAttributeSelectionSchema) o;
-            return consumerSchema.equals(that.consumerSchema) && producerSchema.equals(that.producerSchema);
-        }
-
-        @Override
-        public int hashCode() {
-            return Objects.hashCode(consumerSchema, producerSchema);
-        }
-    }
-
-    /**
-     * A cache entry key, leveraging _identity_ as the key, because we do interning.
-     * This is a performance optimization.
-     */
-    private static class ExtraAttributesEntry {
-        private final ImmutableAttributes[] candidateAttributeSets;
-        private final ImmutableAttributes requestedAttributes;
-        private final int hashCode;
-
-        private ExtraAttributesEntry(ImmutableAttributes[] candidateAttributeSets, ImmutableAttributes requestedAttributes) {
-            this.candidateAttributeSets = candidateAttributeSets;
-            this.requestedAttributes = requestedAttributes;
-            int hash = Arrays.hashCode(candidateAttributeSets);
-            hash = 31 * hash + requestedAttributes.hashCode();
-            this.hashCode = hash;
-        }
-
-        @Override
-        public boolean equals(Object o) {
-            if (this == o) {
-                return true;
-            }
-            if (o == null || getClass() != o.getClass()) {
-                return false;
-            }
-
-            ExtraAttributesEntry that = (ExtraAttributesEntry) o;
-            if (requestedAttributes != that.requestedAttributes) {
-                return false;
-            }
-            if (candidateAttributeSets.length != that.candidateAttributeSets.length) {
-                return false;
-            }
-            for (int i = 0; i < candidateAttributeSets.length; i++) {
-                if (candidateAttributeSets[i] != that.candidateAttributeSets[i]) {
-                    return false;
-                }
-            }
-            return true;
-        }
-
-        @Override
-        public int hashCode() {
-            return hashCode;
-        }
     }
 
 }
