+++ platforms/software/dependency-management/src/test/groovy/org/gradle/api/internal/attributes/DefaultAttributesSchemaTest.groovy
@@ -22,6 +22,7 @@ import org.gradle.api.attributes.AttributeCompatibilityRule
 import org.gradle.api.attributes.AttributeDisambiguationRule
 import org.gradle.api.attributes.CompatibilityCheckDetails
 import org.gradle.api.attributes.MultipleCandidatesDetails
+import org.gradle.api.internal.attributes.matching.DefaultAttributeSelectionSchema
 import org.gradle.util.AttributeTestUtil
 import org.gradle.util.SnapshotTestUtil
 import org.gradle.util.TestUtil
@@ -69,59 +70,6 @@ class DefaultAttributesSchemaTest extends Specification {
         then:
         def e = thrown(IllegalArgumentException)
         e.message == 'Unable to find matching strategy for flavor'
-    }
-
-    def "treats equal values as compatible when no rules defined"() {
-        given:
-        def attribute = Attribute.of(String)
-        schema.attribute(attribute)
-
-        expect:
-        !schema.matcher().selectionSchema.matchValue(attribute, "a", "b")
-        schema.matcher().selectionSchema.matchValue(attribute, "a", "a")
-
-        !schema.matcher().isMatchingValue(attribute, "a", "b")
-        schema.matcher().isMatchingValue(attribute, "a", "a")
-    }
-
-    static class DoNothingRule implements AttributeCompatibilityRule<String> {
-        static int count
-
-        @Override
-        void execute(CompatibilityCheckDetails<String> stringCompatibilityCheckDetails) {
-            count++
-        }
-    }
-
-    def "treats equal values as compatible when no rule expresses an opinion"() {
-        given:
-        def attribute = Attribute.of(String)
-        schema.attribute(attribute).compatibilityRules.add(DoNothingRule)
-
-        expect:
-        !schema.matcher().selectionSchema.matchValue(attribute, "a", "b")
-        schema.matcher().selectionSchema.matchValue(attribute, "a", "a")
-
-        !schema.matcher().isMatchingValue(attribute, "a", "b")
-        schema.matcher().isMatchingValue(attribute, "a", "a")
-    }
-
-    static class BrokenRule implements AttributeCompatibilityRule<String> {
-        @Override
-        void execute(CompatibilityCheckDetails<String> stringCompatibilityCheckDetails) {
-            throw new RuntimeException()
-        }
-    }
-
-    def "short-circuits evaluation when values are equal"() {
-        given:
-        def attribute = Attribute.of(String)
-        schema.attribute(attribute).compatibilityRules.add(BrokenRule)
-
-        expect:
-        schema.matcher().selectionSchema.matchValue(attribute, "a", "a")
-
-        schema.matcher().isMatchingValue(attribute, "a", "a")
     }
 
     def "strategy is per attribute"() {
@@ -176,129 +124,6 @@ class DefaultAttributesSchemaTest extends Specification {
         }
     }
 
-    def "compatibility rules can mark values as compatible"() {
-        def attr = Attribute.of(Flavor)
-
-        given:
-        schema.attribute(attr).compatibilityRules.add(CustomCompatibilityRule)
-
-        def value1 = flavor('value')
-        def value2 = flavor('otherValue')
-
-        expect:
-        schema.matcher().selectionSchema.matchValue(attr, value1, value2)
-        !schema.matcher().selectionSchema.matchValue(attr, value2, value1)
-
-        schema.matcher().isMatchingValue(attr, value2, value1)
-        !schema.matcher().isMatchingValue(attr, value1, value2)
-    }
-
-    static class IncompatibleStringsRule implements AttributeCompatibilityRule<String> {
-        @Override
-        void execute(CompatibilityCheckDetails<String> details) {
-            details.incompatible()
-        }
-    }
-
-    def "compatibility rules can mark values as incompatible and short-circuit evaluation"() {
-        def attr = Attribute.of(String)
-
-        given:
-        schema.attribute(attr).compatibilityRules.add(IncompatibleStringsRule)
-        schema.attribute(attr).compatibilityRules.add(BrokenRule)
-
-        expect:
-        !schema.matcher().selectionSchema.matchValue(attr, "a", "b")
-
-        !schema.matcher().isMatchingValue(attr, "a", "b")
-    }
-
-    def "selects requested value when it is one of the candidate values and no rules defined"() {
-        def attr = Attribute.of(String)
-
-        given:
-        schema.attribute(attr)
-        def candidates = ["foo", "bar"] as Set
-
-        when:
-        def best = schema.matcher().selectionSchema.disambiguate(attr, "bar", candidates)
-
-        then:
-        best == ["bar"] as Set
-    }
-
-    static class DoNothingSelectionRule implements AttributeDisambiguationRule<String> {
-        @Override
-        void execute(MultipleCandidatesDetails<String> stringMultipleCandidatesDetails) {
-        }
-    }
-
-    def "selects requested value when it is one of the candidate values and no rule expresses an opinion"() {
-        def attr = Attribute.of(String)
-
-        given:
-        schema.attribute(attr).disambiguationRules.add(DoNothingSelectionRule)
-        def candidates = ["foo", "bar"] as Set
-
-        when:
-        def best = schema.matcher().selectionSchema.disambiguate(attr, "bar", candidates)
-
-        then:
-        best == ["bar"] as Set
-    }
-
-    def "returns null when no disambiguation rules and requested is not one of the candidate values"() {
-        def attr = Attribute.of(String)
-
-        given:
-        schema.attribute(attr)
-        def candidates = ["foo", "bar"] as Set
-
-        when:
-        def best = schema.matcher().selectionSchema.disambiguate(attr, "other", candidates)
-
-        then:
-        best == null
-    }
-
-    def "returns null when no rule expresses an opinion and requested is not one of the candidate values"() {
-        def attr = Attribute.of(String)
-
-        given:
-        schema.attribute(attr).disambiguationRules.add(DoNothingSelectionRule)
-        def candidates = ["foo", "bar"] as Set
-
-        when:
-        def best = schema.matcher().selectionSchema.disambiguate(attr, "other", candidates)
-
-        then:
-        best == null
-    }
-
-    def "custom rule can select best match"() {
-        def attr = Attribute.of(Flavor)
-
-        given:
-        schema.attribute(attr).disambiguationRules.add(CustomSelectionRule)
-
-        def value1 = flavor('value1')
-        def value2 = flavor('value2')
-
-        def candidates = [value1, value2] as Set
-
-        when:
-        def best = schema.matcher().selectionSchema.disambiguate(attr, flavor('requested'), candidates)
-
-        then:
-        best == [value1] as Set
-
-        when:
-        best = schema.matcher().selectionSchema.disambiguate(attr, value2, candidates)
-
-        then:
-        best == [value1] as Set
-    }
-
     def "merging creates schema with additional attributes defined by producer"() {
         def producer = new DefaultAttributesSchema(TestUtil.instantiatorFactory(), SnapshotTestUtil.isolatableFactory())
 
@@ -312,9 +137,9 @@ class DefaultAttributesSchemaTest extends Specification {
         producer.attribute(attr3)
 
         expect:
-        schema.withProducer(producer).selectionSchema.hasAttribute(attr1)
-        schema.withProducer(producer).selectionSchema.hasAttribute(attr2)
-        schema.withProducer(producer).selectionSchema.hasAttribute(attr3)
+        new DefaultAttributeSelectionSchema(schema, producer).hasAttribute(attr1)
+        new DefaultAttributeSelectionSchema(schema, producer).hasAttribute(attr2)
+        new DefaultAttributeSelectionSchema(schema, producer).hasAttribute(attr3)
     }
 
     def "uses the producers compatibility rules when the consumer does not express an opinion"() {
@@ -326,7 +151,7 @@ class DefaultAttributesSchemaTest extends Specification {
         producer.attribute(attr).compatibilityRules.add(CustomCompatibilityRule)
 
         expect:
-        schema.withProducer(producer).selectionSchema.matchValue(attr, flavor('value'), flavor('otherValue'))
+        schema.withProducer(producer).isMatchingValue(attr, flavor('otherValue'), flavor('value'))
     }
 
     def "uses the producers disambiguation rules when the consumer does not express an opinion"() {
@@ -343,7 +168,7 @@ class DefaultAttributesSchemaTest extends Specification {
         def candidates = [value1, value2] as Set
 
         when:
-        def best = schema.withProducer(producer).selectionSchema.disambiguate(attr, flavor('requested'), candidates)
+        def best = new DefaultAttributeSelectionSchema(schema, producer).disambiguate(attr, flavor('requested'), candidates)
 
         then:
         best == [value1] as Set
@@ -358,7 +183,7 @@ class DefaultAttributesSchemaTest extends Specification {
         producer.attribute(attr).compatibilityRules.add(FailingCompatibilityRule)
 
         expect:
-        schema.withProducer(producer).selectionSchema.matchValue(attr, flavor('value'), flavor('otherValue'))
+        schema.withProducer(producer).isMatchingValue(attr, flavor('otherValue'), flavor('value'))
     }
 
     def "uses the consumer's disambiguation rules when both the consumer and producer express an opinion"() {
@@ -375,7 +200,7 @@ class DefaultAttributesSchemaTest extends Specification {
         def candidates = [value1, value2] as Set
 
         when:
-        def best = schema.withProducer(producer).selectionSchema.disambiguate(attr, flavor('requested'), candidates)
+        def best = new DefaultAttributeSelectionSchema(schema, producer).disambiguate(attr, flavor('requested'), candidates)
 
         then:
         best == [value1] as Set
@@ -428,7 +253,7 @@ class DefaultAttributesSchemaTest extends Specification {
         )
 
         then:
-        def result = schema.withProducer(producer).selectionSchema.orderByPrecedence(requested.keySet())
+        def result = new DefaultAttributeSelectionSchema(schema, producer).orderByPrecedence(requested.keySet())
         result.sortedOrder == [2, 1, 0]
         result.unsortedOrder as List == [3]
     }
@@ -457,68 +282,9 @@ class DefaultAttributesSchemaTest extends Specification {
         )
 
         then:
-        def result = schema.withProducer(producer).selectionSchema.orderByPrecedence(requested.keySet())
+        def result = new DefaultAttributeSelectionSchema(schema, producer).orderByPrecedence(requested.keySet())
         result.sortedOrder == [1, 0, 2]
         result.unsortedOrder as List == [3]
-    }
-
-    def "precedence order is honored"() {
-        def x = Attribute.of("x", String)
-        def a = Attribute.of("a", Flavor)
-        def b = Attribute.of("b", String)
-        def c = Attribute.of("c", ConcreteNamed)
-        schema.attributeDisambiguationPrecedence(a, b, c)
-
-        def requested = AttributeTestUtil.attributesTyped(
-                // attribute that doesn't have a precedence
-                (x): "x",
-                // attribute that has a lower precedence than the next one
-                (c): AttributeTestUtil.named(ConcreteNamed, "c"),
-                // attribute with the highest precedence
-                (a): flavor("a"),
-                // attribute that doesn't have a precedence
-                (Attribute.of("z", String)): "z"
-        )
-        expect:
-        def result = schema.matcher().selectionSchema.orderByPrecedence(requested.keySet())
-        result.sortedOrder == [2, 1]
-        result.unsortedOrder as List == [0, 3]
-    }
-
-    def "requested attributes are not sorted when there is no attribute precedence"() {
-        def x = Attribute.of("x", String)
-        def a = Attribute.of("a", Flavor)
-        def c = Attribute.of("c", ConcreteNamed)
-
-        def requested = AttributeTestUtil.attributesTyped(
-                (x): "x",
-                (c): AttributeTestUtil.named(ConcreteNamed, "c"),
-                (a): flavor("a"),
-                (Attribute.of("z", String)): "z"
-        )
-        expect:
-        def result = schema.matcher().selectionSchema.orderByPrecedence(requested.keySet())
-        result.sortedOrder == []
-        result.unsortedOrder as List == [0, 1, 2, 3]
-    }
-
-    def "requested attributes are not sorted when there is a different set of attributes used for precedence"() {
-        def x = Attribute.of("x", String)
-        def a = Attribute.of("a", Flavor)
-        def c = Attribute.of("c", ConcreteNamed)
-
-        schema.attributeDisambiguationPrecedence(Attribute.of("notA", Flavor), Attribute.of("notB", String), Attribute.of("notC", ConcreteNamed))
-
-        def requested = AttributeTestUtil.attributesTyped(
-                (x): "x",
-                (c): AttributeTestUtil.named(ConcreteNamed, "c"),
-                (a): flavor("a"),
-                (Attribute.of("z", String)): "z"
-        )
-        expect:
-        def result = schema.matcher().selectionSchema.orderByPrecedence(requested.keySet())
-        result.sortedOrder == []
-        result.unsortedOrder as List == [0, 1, 2, 3]
     }
 
     static interface Flavor extends Named {}
